+++ src/app/meter/meters-list/meters-list.component.ts
@@ -7,6 +7,8 @@ import Meter from '../../models/Meter';
 import {UserService} from '../../user/user.service';
 import {Router} from '@angular/router';
 import {Subscription} from 'rxjs';
+import {MatDialog, MatDialogRef} from '@angular/material';
+import {ConfirmComponent} from '../../confirm/confirm.component';
 
 @Component({
   selector: 'app-meters-list',
@@ -22,7 +24,9 @@ export class MetersListComponent implements OnInit, OnChanges, OnDestroy {
   @Input() private categoryId: number;
   meters: Meter[];
 
-  constructor(private meterService: MeterService,private userService: UserService, private router: Router) {
+  dialogRef: MatDialogRef<ConfirmComponent, any>;
+
+  constructor(private meterService: MeterService, private userService: UserService, private router: Router, private dialog: MatDialog) {
   }
 
   ngOnInit() {
@@ -59,21 +63,34 @@ export class MetersListComponent implements OnInit, OnChanges, OnDestroy {
 
   deleteMeter(meterId: number, event) {
     event.stopPropagation();
+
     this.getMeterByIdSubscription = this.meterService.getMeterById(meterId)
       .subscribe((response: HttpResponse<any>) => {
         if (response) {
           tokenSetter(response);
           const meter = response.body;
 
-          this.deleteMeterSubscription = this.meterService.deleteMeter(meter)
-            .subscribe((deleteResp: HttpResponse<any>) => {
-              if (deleteResp) {
-                this.router.navigateByUrl('/home', {skipLocationChange: true}).then(() =>
-                  this.router.navigate(['/category/user/' + this.userService.currentUser.id]));
-              }
-            }, (appError: AppError) => {
-              throw appError;
-            });
+          this.dialogRef = this.dialog.open(ConfirmComponent, {
+            disableClose: true,
+            autoFocus: false
+          });
+          this.dialogRef.componentInstance.confirmMessage = `Ви дійсно хоче видалити лічильник з назвою "${meter.name}"?`;
+          this.dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+
+              this.deleteMeterSubscription = this.meterService.deleteMeter(meter)
+                .subscribe((deleteResp: HttpResponse<any>) => {
+                  if (deleteResp) {
+                    this.router.navigateByUrl('/home', {skipLocationChange: true}).then(() =>
+                      this.router.navigate(['/category/user/' + this.categoryId]));
+                  }
+                }, (appError: AppError) => {
+                  throw appError;
+                });
+
+            }
+            this.dialogRef = null;
+          });
 
         }
       }, (appError: AppError) => {
